backend/image_similarity/requirements.txt
scikit-image==0.2*
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==2.2.2
google-cloud-storage==2.14.0
h11==0.14.0
huggingface-hub==0.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.3
outcome==1.2.0
packaging==21.3
Pillow==10.2.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.5.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.22.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0